[fix] move type, hook
Hooks와 Types는 컴포넌트가 아니므로 atoms -> src로 이동
경로 수정하면서 enum 반영 누락됐던 부분도 같이 수정
